Replace Position2 property with ScalarBarLength
A separate ScalarBarLength property is much more intuitive than the
Position2 property that required the right element to be set depending
on the orientation. Also, Position2 is a misnomer - it is a relative
position, but really it specifies the size of the scalar bar.

Unfortunately, the Position2 attribute is deep down in the class
hierarchy, so we are stuck with it when interacting with the scalar
bar widget as it stands. This commit adds necessary translation code
that synchronizes the ScalarBarLength property with the Position2
property.

+++ ParaViewCore/VTKExtensions/Rendering/vtkContext2DScalarBarActor.h
@@ -83,11 +83,14 @@ public:
   //@{
   /**
    * Set the scalar bar length. When the orientation is VTK_ORIENT_VERTICAL,
-   * this sets the scalar bar  height. When the orientation is
-   * VTK_ORIENT_HORIZONTAL, this sets the scalar bar width. Specified in pixels.
+   * this sets the scalar bar height. When the orientation is
+   * VTK_ORIENT_HORIZONTAL, this sets the scalar bar width. Specified in
+   * normalized viewport coordinates, meaning the value is the fractional span
+   * of the viewport's width or height in the range [0, 1], depending on
+   * orientation.
    */
-  vtkSetMacro(ScalarBarLength, int);
-  vtkGetMacro(ScalarBarLength, int);
+  vtkSetMacro(ScalarBarLength, double);
+  vtkGetMacro(ScalarBarLength, double);
   //@}
 
   //@{
@@ -214,7 +217,7 @@ private:
   /**
   * Length of the color bar.
   */
-  int ScalarBarLength;
+  double ScalarBarLength;
 
   int AutomaticLabelFormat;
